Add base activity that just applies theme.
